move database and jukebox crates into trollstov
