--- crates/database/Cargo.toml
@@ -1,23 +1,0 @@
-[package]
-name = "database"
-version.workspace = true
-edition.workspace = true
-rust-version.workspace = true
-homepage.workspace = true
-repository.workspace = true
-license.workspace = true
-
-[dependencies]
-fastrand = { workspace = true }
-ignore = { workspace = true }
-indexmap = { workspace = true }
-lofty = { workspace = true }
-nucleo-matcher = { workspace = true }
-serde = { workspace = true }
-serde_repr = { workspace = true }
-utils = { workspace = true }
-walkdir = { workspace = true }
-
-[features]
-default = []
-opus = []
