--- crates/database/src/lib.rs
@@ -1,10 +1,0 @@
-mod audio;
-mod database;
-mod matcher;
-mod track;
-
-use matcher::*;
-
-pub use audio::*;
-pub use database::*;
-pub use track::*;
