--- crates/jukebox/Cargo.toml
@@ -1,17 +1,0 @@
-[package]
-name = "jukebox"
-version.workspace = true
-edition.workspace = true
-rust-version.workspace = true
-homepage.workspace = true
-repository.workspace = true
-license.workspace = true
-
-[dependencies]
-database = { workspace = true }
-fastrand = { workspace = true }
-rodio = { workspace = true }
-
-[features]
-default = []
-opus = ["rodio/symphonia-libopus"]
