--- crates/jukebox/src/lib.rs
@@ -1,8 +1,0 @@
-mod jukebox;
-mod player;
-mod queue;
-
-use queue::*;
-
-pub use jukebox::*;
-pub use player::*;
