+++ trollstov/Cargo.toml
@@ -22,20 +22,26 @@ categories = ["audio", "command-line-utilities", "multimedia"]
 clap = { workspace = true }
 clap-cargo = { workspace = true }
 constcat = { workspace = true }
-database = { workspace = true }
 directories = { workspace = true }
+fastrand = { workspace = true }
+ignore = { workspace = true }
 image = { workspace = true }
-jukebox = { workspace = true }
+indexmap = { workspace = true }
+lofty = { workspace = true }
+nucleo-matcher = { workspace = true }
 ratatui = { workspace = true }
 ratatui-image = { workspace = true }
+rodio = { workspace = true }
 seahash = { workspace = true }
 serde = { workspace = true }
+serde_repr = { workspace = true }
 souvlaki = { workspace = true }
 terminal-colorsaurus = { workspace = true }
 toml = { workspace = true }
 unicode-segmentation = { workspace = true }
 unicode-width = { workspace = true }
 utils = { workspace = true }
+walkdir = { workspace = true }
 widgets = { workspace = true }
 
 [target.'cfg(target_os = "windows")'.dependencies.windows]
@@ -49,4 +55,4 @@ features = [
 
 [features]
 default = []
-opus = ["database/opus", "jukebox/opus"]
+opus = ["rodio/symphonia-libopus"]
