+++ trollstov/src/app.rs
@@ -1,7 +1,5 @@
 use std::{path::PathBuf, time::Duration};
 
-use database::{Database, DatabaseEvent, Track};
-use jukebox::{Jukebox, JukeboxEvent};
 use ratatui::{
     CompletedFrame,
     crossterm::event::{Event as CrosstermEvent, KeyCode, KeyEvent, KeyEventKind, KeyModifiers},
@@ -11,7 +9,9 @@ use ratatui_image::{picker::Picker, protocol::StatefulProtocol};
 use widgets::{Shortcut, Shortcuts, TextSegment};
 
 use crate::{
+    database::{Database, DatabaseEvent, Track},
     events::{Event, EventHandler, MediaEvent, MediaPlayback},
+    jukebox::{Jukebox, JukeboxEvent},
     pages::{
         Log, LogsAction, LogsPage, Pages, PlayingPage, Route, SearchAction, SearchPage,
         SettingsPage, TracksPage,
@@ -858,7 +858,7 @@ fn fill_app_shortcuts(shortcuts: &mut Shortcuts, logs: &LogsPage) {
 // https://github.com/ratatui/ratatui-image/blob/master/examples/thread.rs
 fn load_front_cover(path: PathBuf, picker: Picker) -> FrontCoverHandle {
     std::thread::spawn(move || {
-        let front_cover = database::AudioFrontCover::read(&path)?;
+        let front_cover = crate::database::AudioFrontCover::read(&path)?;
 
         let Some((bytes, mime_type)) = front_cover.bytes_and_mime_type() else {
             return Ok(FrontCover::empty());
@@ -872,8 +872,8 @@ fn load_front_cover(path: PathBuf, picker: Picker) -> FrontCoverHandle {
         };
 
         let image_format = match mime_type {
-            database::MimeType::Jpeg => image::ImageFormat::Jpeg,
-            database::MimeType::Png => image::ImageFormat::Png,
+            crate::database::MimeType::Jpeg => image::ImageFormat::Jpeg,
+            crate::database::MimeType::Png => image::ImageFormat::Png,
             _ => {
                 return Err(format!(
                     "Unable to load front cover from \"{}\" due to unsupported or unknown mime type: {}",
