+++ trollstov/src/jukebox.rs
@@ -2,9 +2,13 @@ use std::{collections::HashSet, io::Cursor, time::Duration};
 
 use rodio::decoder::Decoder;
 
-use database::*;
+mod player;
+mod queue;
 
-use crate::{AudioPlayer, PlayQueue};
+pub use player::*;
+use queue::*;
+
+use crate::database::{AudioRating, Database, TrackId};
 
 type AudioDecodeHandle = std::thread::JoinHandle<Result<Decoder<Cursor<Vec<u8>>>, String>>;
