+++ trollstov/src/jukebox/queue.rs
@@ -1,4 +1,4 @@
-use database::TrackId;
+use crate::database::TrackId;
 
 // TODO: Max length? Drain from history.
